pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-math-tutor"
version = "0.1.0"
description = "Local AI math tutor for kids"
requires-python = ">=3.9"
dependencies = [
    "streamlit>=1.35.0",
    "ollama>=0.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "pytest>=8.0.0",
]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
